Move global types to separate file
