+++ packages/browser-data-collector/src/types.ts
@@ -1,25 +1,3 @@
-export type Navigator = typeof navigator & {
-	deviceMemory: number;
-	connection: {
-		effectiveType: EffectiveConnectionType;
-	};
-};
-
-export enum EffectiveConnectionType {
-	'2g',
-	'3g',
-	'4g',
-	'slow-2g',
-}
-
-export type Window = typeof window & {
-	COMMIT_SHA: string;
-	BUILD_TARGET: string;
-	configData: {
-		env_id: string;
-	};
-};
-
 export type ReportData = Map< string, number | string | boolean >;
 
 export type ReportPayload = Record< string, number | string | boolean >;
